Move edit buttons into own file
